pyproject.toml
[project]
name = "raycast_backend"
version = "0.2.0"
description = ""
authors = [
    {name = "Yufei Kang", email = "kou.uhi.x@gmail.com"},
]
dependencies = [
    "fastapi>=0.95.1",
    "openai>=1.3.6",
    "google-genai>=1.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0.2",
    "jsonpath-ng>=1.6.1",
    "anthropic>=0.29.0",
    "click>=8.1.7",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
raycast-proxy = "raycast_proxy.cli:main [cli]"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["raycast_proxy"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
atomic = true
profile = "black"
line_length = 88
skip_gitignore = true

[project.optional-dependencies]
cli = [
    "uvicorn[standard]>=0.30.1",
    "cryptography>=42.0.7",
    "mitmproxy>=11.0.2",
]
dev = [
    "black>=24.1.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
